Extracted blazor logic to BlazorSourcePathResolver
